[flang2] Declare static visited_size as an int; NFCI
The static variable visited_size in tools/flang2/flang2exe/verify.cpp
denotes the number of elements in the reallocatable array visited_ili,
is used frequently in comparison with int-type variables (e.g. ilix),
and does not need to be of type size_t. This fixes a number of
-Wsign-compare warnings.

+++ tools/flang2/flang2exe/verify.cpp
@@ -35,7 +35,7 @@ typedef struct {
 } visit_info;
 
 static visit_info *visited_ili;
-static size_t visited_size; /* Number of elements pointed to by visited_ili. */
+static int visited_size; /* Number of elements pointed to by visited_ili. */
 static epoch_t current_epoch;
 static int walk_depth;
 
@@ -74,7 +74,7 @@ end_walk(VERIFY_LEVEL level)
 static void
 ili_mark_first_visit(int ilix)
 {
-  size_t old_size;
+  int old_size;
   visit_info *v;
   DEBUG_ASSERT(walk_depth > 0, "internal error in verifier: not in a walk");
   DEBUG_ASSERT(0 < ilix, "internal error in verifier itself");
